Pre-rotate the anchor transform
This is the launch clamp side of the fix for the EL-ReStock launch clamp
compatibility issue. The issue was reported in the EL thread and started
out as issues with getting the build cost but after fixing that, had
issues with the clamps extending properly when built at a survey site.
The bulk of the fix for that is (or will be) in EL, but this change is
for handling rotated clamps: it points the tower in the correct (editor)
direction prior to the tower being extended.