Set entryCost to zero on TechHidden parts
A TechHidden part's entryCost isn't included in the price shown on the
tech node's "purchase all parts" button, but clicking the button
actually charges the player for hidden parts nonetheless.  This is a bug
in the game, but it has a simple workaround: set entryCost to zero on
parts that have the TechHidden flag.  (Squad's deprecated parts do this
too.)